recrutor-backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-multipart>=0.0.5
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
alembic>=1.7.3
openai>=0.27.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
PyPDF2>=2.10.0
docx2txt>=0.8.0
pydantic-settings
shap>=0.41.0
lime>=0.2.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
